@@ -3,22 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chec/commerce.js": "^2.8.0",
+    "@chec/commerce.js": "^2.4.0",
     "@emailjs/browser": "^3.6.2",
     "@stripe/react-stripe-js": "^1.7.0",
     "@stripe/stripe-js": "^1.26.0",
-    "@testing-library/jest-dom": "^5.16.3",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
     "framer-motion": "^6.2.9",
-    "node-sass": "^7.0.1",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-feather": "^2.0.9",
     "react-intersection-observer": "^8.34.0",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "striptags": "^3.2.0",
+    "striptags": "^3.2.0"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.69.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
