@@ -1,11 +1,12 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import { HashRouter } from "react-router-dom";
+import { createRoot } from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 import App from "./App";
 
-ReactDOM.render(
-  <HashRouter>
+const root = createRoot(document.getElementById("root"));
+root.render(
+  <BrowserRouter>
     <App />
-  </HashRouter>,
-  document.getElementById("root")
+  </BrowserRouter>
 );
+
